Fix: reworked rename() and save(), to fix various of corner-cases
Those two functions are now also merged, which is easier to use
from a user perspective.

+++ truewiki/views/edit.py
@@ -7,7 +7,10 @@ from . import (
     error,
     source,
 )
-from .. import singleton
+from .. import (
+    metadata,
+    singleton,
+)
 from ..metadata import page_changed
 from ..wiki_page import WikiPage
 
@@ -17,7 +20,7 @@ def _check_illegal_names(user, page: str, new_page: str) -> web.Response:
         return error.view(
             user,
             page,
-            f"Page '{new_page}' contain '..' and/or starts/ends with a space and/or dot, which is not allowed.",
+            f"Page name '{new_page}' contain '..' and/or starts/ends with a space and/or dot, which is not allowed.",
             status=401,
         )
 
@@ -26,7 +29,7 @@ def _check_illegal_names(user, page: str, new_page: str) -> web.Response:
             return error.view(
                 user,
                 page,
-                f'Page "{new_page}" contains '
+                f'Page name "{new_page}" contains '
                 '<a href="https://en.wikipedia.org/wiki/Control_character" target="_new">Control Characters</a>, '
                 "which is not allowed.",
                 status=401,
@@ -35,21 +38,35 @@ def _check_illegal_names(user, page: str, new_page: str) -> web.Response:
     return None
 
 
-def rename(user, old_page: str, new_page: str) -> web.Response:
-    if new_page == old_page:
-        return web.HTTPFound(f"/{old_page}")
+def save(user, old_page: str, new_page: str, content: str) -> web.Response:
+    wiki_page = WikiPage(old_page)
+    if not wiki_page.page_is_valid(old_page):
+        return error.view(user, old_page, "Error 404 - File not found")
+    if not wiki_page.page_is_valid(new_page):
+        return error.view(
+            user,
+            new_page,
+            f'Page name "{new_page}" is not a valid name; does it start with a language code?',
+        )
 
     response = _check_illegal_names(user, old_page, new_page)
-    if response:
+    if response is not None:
         return response
 
-    wiki_page = WikiPage(old_page)
+    # If the old page doesn't exist, this creates a page. But it is possible
+    # that this is done from an old page with a different name. Make sure the
+    # new page always wins in this case.
     if not wiki_page.page_exists(old_page):
-        return error.view(user, old_page, "Error 404 - File not found")
+        create_new = True
+        old_page = new_page
+    else:
+        create_new = False
+
+    old_filename = wiki_page.page_ondisk_name(old_page)
 
     # If we are not renaming to the same page (in lower-case), ensure we are
     # not renaming to a page similar but with different casing.
-    if old_page.lower() != new_page.lower():
+    if old_page.lower() != new_page.lower() or create_new:
         if wiki_page.page_exists(new_page):
             return error.view(
                 user,
@@ -65,63 +82,31 @@ def rename(user, old_page: str, new_page: str) -> web.Response:
                 f'Page name "{new_page}" conflicts with "{correct_page}", which already exists.',
             )
 
-    # Read the old file from disk.
-    old_filename = wiki_page.page_ondisk_name(old_page)
-    with open(f"{singleton.STORAGE.folder}/{old_filename}") as fp:
-        body = fp.read()
+        # Validate that if the page exists, we are not renaming it while there are
+        # other pages depending on us
+        if wiki_page.page_exists(old_page) and metadata.TEMPLATES[old_filename[: -len(".mediawiki")]]:
+            return error.view(user, old_page, "Cannot rename page as other pages depend on it.")
+
+    changed = []
+
+    if old_page != new_page:
+        # Remove the old file.
+        os.unlink(f"{singleton.STORAGE.folder}/{old_filename}")
+        changed.append(old_filename[: -len(".mediawiki")])
 
     new_filename = wiki_page.page_ondisk_name(new_page)
     new_dirname = "/".join(new_filename.split("/")[:-1])
-
-    # Remove the old file.
-    os.unlink(f"{singleton.STORAGE.folder}/{old_filename}")
 
     # Make sure the folder exists.
     os.makedirs(f"{singleton.STORAGE.folder}/{new_dirname}", exist_ok=True)
     # Write the new source.
     with open(f"{singleton.STORAGE.folder}/{new_filename}", "w") as fp:
-        fp.write(body)
+        fp.write(content)
 
-    page_changed(
-        [
-            old_filename[: -len(".mediawiki")],
-            new_filename[: -len(".mediawiki")],
-        ]
-    )
+    changed.append(new_filename[: -len(".mediawiki")])
+    page_changed(changed)
 
     return web.HTTPFound(f"/{new_page}")
-
-
-def save(user, page: str, change: str) -> web.Response:
-    wiki_page = WikiPage(page)
-    if not wiki_page.page_is_valid(page):
-        return error.view(user, page, "Error 404 - File not found")
-
-    response = _check_illegal_names(user, page, page)
-    if response:
-        return response
-
-    # If there is a difference in case, nicely point this out to users.
-    correct_page = wiki_page.page_get_correct_case(page)
-    if correct_page != page:
-        return error.view(
-            user,
-            page,
-            f'Page name "{page}" conflicts with "{correct_page}". Did you mean to edit [[{correct_page}]]?',
-        )
-
-    filename = wiki_page.page_ondisk_name(page)
-    dirname = "/".join(filename.split("/")[:-1])
-
-    # Make sure the folder exists.
-    os.makedirs(f"{singleton.STORAGE.folder}/{dirname}", exist_ok=True)
-    # Write the new source.
-    with open(f"{singleton.STORAGE.folder}/{filename}", "w") as fp:
-        fp.write(change)
-
-    page_changed([filename[: -len(".mediawiki")]])
-
-    return web.HTTPFound(f"/{page}")
 
 
 def view(user, page: str) -> web.Response:
